refactor(WinningLotto, Game): splitWinningLottoNumber 위치 변경
기존 `Game`에 존재하던 `splitWinningLottoNumbers`가 `Game`에 존재할 이유가 없다고 판단하여 `WinningLotto`로 위치를 옮김.
